[BEAM-14489] Remove non-SDF version of TextIO. (#17712)
