+++ sdks/go/pkg/beam/io/textio/textio.go
@@ -25,18 +25,23 @@ import (
 	"strings"
 
 	"github.com/apache/beam/sdks/v2/go/pkg/beam"
+	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/sdf"
+	"github.com/apache/beam/sdks/v2/go/pkg/beam/internal/errors"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/io/filesystem"
+	"github.com/apache/beam/sdks/v2/go/pkg/beam/io/rtrackers/offsetrange"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/log"
 )
 
 func init() {
+	beam.RegisterType(reflect.TypeOf((*readFn)(nil)).Elem())
+	beam.RegisterFunction(sizeFn)
 	beam.RegisterType(reflect.TypeOf((*writeFileFn)(nil)).Elem())
-	beam.RegisterFunction(readFn)
 	beam.RegisterFunction(expandFn)
 }
 
-// Read reads a set of file and returns the lines as a PCollection<string>. The
-// newlines are not part of the lines.
+// Read reads a set of files indicated by the glob pattern and returns
+// the lines as a PCollection<string>.
+// The newlines are not part of the lines.
 func Read(s beam.Scope, glob string) beam.PCollection {
 	s = s.Scope("textio.Read")
 
@@ -49,15 +54,39 @@ func Read(s beam.Scope, glob string) beam.PCollection {
 // PCollection<string>. The newlines are not part of the lines.
 func ReadAll(s beam.Scope, col beam.PCollection) beam.PCollection {
 	s = s.Scope("textio.ReadAll")
-
 	return read(s, col)
 }
 
-func read(s beam.Scope, col beam.PCollection) beam.PCollection {
-	files := beam.ParDo(s, expandFn, col)
-	return beam.ParDo(s, readFn, files)
+// ReadSdf is a variation of Read implemented via SplittableDoFn. This should
+// result in increased performance with runners that support splitting.
+//
+// Deprecated: Use Read instead, which has been migrated to use this SDF implementation.
+func ReadSdf(s beam.Scope, glob string) beam.PCollection {
+	s = s.Scope("textio.ReadSdf")
+
+	filesystem.ValidateScheme(glob)
+	return read(s, beam.Create(s, glob))
 }
 
+// ReadAllSdf is a variation of ReadAll implemented via SplittableDoFn. This
+// should result in increased performance with runners that support splitting.
+//
+// Deprecated: Use ReadAll instead, which has been migrated to use this SDF implementation.
+func ReadAllSdf(s beam.Scope, col beam.PCollection) beam.PCollection {
+	s = s.Scope("textio.ReadAllSdf")
+	return read(s, col)
+}
+
+// read takes a PCollection of globs and returns a PCollection of lines from
+// all files in those globs. Uses an SDF to allow splitting reads of files
+// into separate bundles.
+func read(s beam.Scope, col beam.PCollection) beam.PCollection {
+	files := beam.ParDo(s, expandFn, col)
+	sized := beam.ParDo(s, sizeFn, files)
+	return beam.ParDo(s, &readFn{}, sized)
+}
+
+// expandFn expands a glob pattern into all matching file names.
 func expandFn(ctx context.Context, glob string, emit func(string)) error {
 	if strings.TrimSpace(glob) == "" {
 		return nil // ignore empty string elements here
@@ -79,7 +108,82 @@ func expandFn(ctx context.Context, glob string, emit func(string)) error {
 	return nil
 }
 
-func readFn(ctx context.Context, filename string, emit func(string)) error {
+// sizeFn pairs a filename with the size of that file in bytes.
+// TODO(BEAM-11109): Once CreateInitialRestriction supports Context params and
+// error return values, this can be done in readSdfFn.CreateInitialRestriction.
+func sizeFn(ctx context.Context, filename string) (string, int64, error) {
+	fs, err := filesystem.New(ctx, filename)
+	if err != nil {
+		return "", -1, err
+	}
+	defer fs.Close()
+
+	size, err := fs.Size(ctx, filename)
+	if err != nil {
+		return "", -1, err
+	}
+	return filename, size, nil
+}
+
+// readFn reads individual lines from a text file, given a filename and a
+// size in bytes for that file. Implemented as an SDF to allow splitting
+// within a file.
+type readFn struct {
+}
+
+// CreateInitialRestriction creates an offset range restriction representing
+// the file, using the paired size rather than fetching the file's size.
+func (fn *readFn) CreateInitialRestriction(_ string, size int64) offsetrange.Restriction {
+	return offsetrange.Restriction{
+		Start: 0,
+		End:   size,
+	}
+}
+
+const (
+	// blockSize is the desired size of each block for initial splits.
+	blockSize int64 = 64 * 1024 * 1024 // 64 MB
+	// tooSmall is the size limit for a block. If the last block is smaller than
+	// this, it gets merged with the previous block.
+	tooSmall = blockSize / 4
+)
+
+// SplitRestriction splits each file restriction into blocks of a predeterined
+// size, with some checks to avoid having small remainders.
+func (fn *readFn) SplitRestriction(_ string, _ int64, rest offsetrange.Restriction) []offsetrange.Restriction {
+	splits := rest.SizedSplits(blockSize)
+	numSplits := len(splits)
+	if numSplits > 1 {
+		last := splits[numSplits-1]
+		if last.End-last.Start <= tooSmall {
+			// Last restriction is too small, so merge it with previous one.
+			splits[numSplits-2].End = last.End
+			splits = splits[:numSplits-1]
+		}
+	}
+	return splits
+}
+
+// Size returns the size of each restriction as its range.
+func (fn *readFn) RestrictionSize(_ string, _ int64, rest offsetrange.Restriction) float64 {
+	return rest.Size()
+}
+
+// CreateTracker creates sdf.LockRTrackers wrapping offsetRange.Trackers for
+// each restriction.
+func (fn *readFn) CreateTracker(rest offsetrange.Restriction) *sdf.LockRTracker {
+	return sdf.NewLockRTracker(offsetrange.NewTracker(rest))
+}
+
+// ProcessElement outputs all lines in the file that begin within the paired
+// restriction.
+//
+// Note that restrictions may not align perfectly with lines. So lines can begin
+// before the restriction and end within it (those are ignored), and lines can
+// begin within the restriction and past the restriction (those are entirely
+// output, including the portion outside the restriction). In some cases a
+// valid restriction might not output any lines.
+func (fn *readFn) ProcessElement(ctx context.Context, rt *sdf.LockRTracker, filename string, _ int64, emit func(string)) error {
 	log.Infof(ctx, "Reading from %v", filename)
 
 	fs, err := filesystem.New(ctx, filename)
@@ -95,18 +199,53 @@ func readFn(ctx context.Context, filename string, emit func(string)) error {
 	defer fd.Close()
 
 	rd := bufio.NewReader(fd)
-	for {
+
+	i := rt.GetRestriction().(offsetrange.Restriction).Start
+	if i > 0 {
+		// If restriction's starts after 0, we cannot assume a new line starts
+		// at the beginning of the restriction, so we must search for the first
+		// line beginning at or after restriction.Start. This is done by
+		// scanning to the byte just before the restriction and then reading
+		// until the next newline, leaving the reader at the start of a new
+		// line past restriction.Start.
+		i -= 1
+		n, err := rd.Discard(int(i)) // Scan to just before restriction.
+		if err == io.EOF {
+			return errors.Errorf("TextIO restriction lies outside the file being read. "+
+				"Restriction begins at %v bytes, but file is only %v bytes.", i+1, n)
+		}
+		if err != nil {
+			return err
+		}
+		line, err := rd.ReadString('\n') // Read until the first line within the restriction.
+		if err == io.EOF {
+			// No lines start in the restriction but it's still valid, so
+			// finish claiming before returning to avoid errors.
+			rt.TryClaim(rt.GetRestriction().(offsetrange.Restriction).End)
+			return nil
+		}
+		if err != nil {
+			return err
+		}
+		i += int64(len(line))
+	}
+
+	// Claim each line until we claim a line outside the restriction.
+	for rt.TryClaim(i) {
 		line, err := rd.ReadString('\n')
 		if err == io.EOF {
 			if len(line) != 0 {
 				emit(strings.TrimSuffix(line, "\n"))
 			}
+			// Finish claiming restriction before breaking to avoid errors.
+			rt.TryClaim(rt.GetRestriction().(offsetrange.Restriction).End)
 			break
 		}
 		if err != nil {
 			return err
 		}
 		emit(strings.TrimSuffix(line, "\n"))
+		i += int64(len(line))
 	}
 	return nil
 }
