+++ sdks/go/pkg/beam/io/textio/textio_test.go
@@ -30,23 +30,6 @@ import (
 
 const testFilePath = "../../../../data/textio_test.txt"
 
-func TestReadFn(t *testing.T) {
-	receivedLines := []string{}
-	getLines := func(line string) {
-		receivedLines = append(receivedLines, line)
-	}
-
-	err := readFn(context.Background(), testFilePath, getLines)
-	if err != nil {
-		t.Fatalf("failed with %v", err)
-	}
-	want := 1
-	if len(receivedLines) != 1 {
-		t.Fatalf("received %v lines, want %v", len(receivedLines), want)
-	}
-
-}
-
 func TestRead(t *testing.T) {
 	p, s := beam.NewPipelineWithRoot()
 	lines := Read(s, testFilePath)
@@ -105,4 +88,28 @@ func TestImmediate(t *testing.T) {
 	passert.Count(s, lines, "NumLines", 2)
 
 	ptest.RunAndValidate(t, p)
+}
+
+// TestReadSdf tests that readSdf successfully reads a test text file, and
+// outputs the correct number of lines for it, even for an exceedingly long
+// line.
+func TestReadSdf(t *testing.T) {
+	p, s := beam.NewPipelineWithRoot()
+	lines := ReadSdf(s, testFilePath)
+	passert.Count(s, lines, "NumLines", 1)
+
+	if _, err := beam.Run(context.Background(), "direct", p); err != nil {
+		t.Fatalf("Failed to execute job: %v", err)
+	}
+}
+
+func TestReadAllSdf(t *testing.T) {
+	p, s := beam.NewPipelineWithRoot()
+	files := beam.Create(s, testFilePath)
+	lines := ReadAllSdf(s, files)
+	passert.Count(s, lines, "NumLines", 1)
+
+	if _, err := beam.Run(context.Background(), "direct", p); err != nil {
+		t.Fatalf("Failed to execute job: %v", err)
+	}
 }
